Be consistent about mipmap count data types: - Use size_t when it is a scalar (Image, Texture::getNumMipmaps() ) - Use int when it is a request which defaults to -1 which means use default (TextureManager::create*) - Add MIP_UNLIMITED as an arbitrarily high number signifying "as much mipmaps as possible" - Never use short, that datatype makes no sense in this context

+++ trunk/OgreMain/src/OgreTexture.cpp
@@ -87,7 +87,7 @@ namespace Ogre {
 		return mWidth * mHeight * mDepth * mFinalBpp;
 	}
 	//--------------------------------------------------------------------------
-	int Texture::getNumFaces(void) const
+	size_t Texture::getNumFaces(void) const
 	{
 		return getTextureType() == TEX_TYPE_CUBE_MAP ? 6 : 1;
 	}
